Resolve volume mount conflicts
Because the volume is mounted from two containers, the contents may be overwritten.
In my testing it's intermittent, but leads to a failure in starting via compose when it does.
This way we ensure that no matter what happens, the defaults file _or_ target file is in place
for either enrichment table before vector attempts to boot.