T-9053: adjust `clone3` seccomp profile
There was a mistake in the profile - the `clone3` action was explicitly disallowed, except under `CAP_SYS_ADMIN`.
This is generally a reasonable protection, except in cases where we tightly control the workload, like in the collector's case.

This patch removes the explicit block, which should allow the `clone3` calls we need.